[Spike] Refactor discord messaging (#93)
Extract discord messaging to own function
