+++ libs/service/discord/src/index.ts
@@ -3,3 +3,4 @@ export * from "./getDiscordBot";
 export * from "./getDiscordUpdate";
 export * from "./getProposalFields";
 export * from "./getVoteFields";
+export * from "./handleDiscordMessaging";
